+++ test/docker_commander_containers_test.dart
@@ -9,8 +9,10 @@ import 'logger_config.dart';
 
 final _log = Logger('docker_commander/test');
 
-void main() {
+Future<void> main() async {
   configureLogger();
+
+  var dockerRunning = await DockerHost.isDaemonRunning(DockerHostLocal());
 
   group('DockerContainerConfig', () {
     late DockerCommander dockerCommander;
@@ -350,5 +352,5 @@ void main() {
 
       await dockerCommander.removeNetwork(network);
     });
-  });
+  }, skip: !dockerRunning);
 }